requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-crispy-forms==1.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pkg_resources==0.0.0
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
